+++ resources/webix_widgets/import.js
@@ -7,6 +7,7 @@ import "./info_datatable.js";
 import "./devices_tree.js";
 import "./device_view_panel.js";
 import "./plot.js";
+import "./scalar_view.js";
 import "./tango_host_view.js";
 import "./device_view.js";
 import "./device_properties_view.js";
@@ -14,6 +15,7 @@ import "./device_polling_view.js";
 import "./device_events_view.js";
 import "./device_attributes_config_view.js";
 import "./device_logging_view.js";
+import "./attrs_monitor_toolbar.js"
 import "./attrs_monitor_view.js";
 import "./device_monitor_view.js";
 import "./scripting_console.js";
